% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cov19datasets}
\alias{cov19datasets}
\alias{cov19state}
\alias{cov19county}
\title{US COVID-19 datasets}
\format{
\code{cov19state}:  Data frame with 5 columns
\describe{
\item{date}{Date of the report}
\item{state}{State reporting}
\item{fips}{FIPS code for the state}
\item{cases}{Cumulative number of COVID-19 cases reported in the state.}
}

\code{cov19county}:  Data frame with 6 columns
\describe{
\item{date}{Date of the report}
\item{county}{County (or county-equivalent) reporting}
\item{state}{State reporting}
\item{fips}{FIPS code for the county}
\item{cases}{Cumulative number of COVID-19 cases reported in the county.}
}
}
\usage{
cov19state

cov19county
}
\description{
Data on COVID-19 cases and deaths by US state and county, going back to 
2020-01-21.  Full documentation of the dataset is available at
https://github.com/nytimes/covid-19-data/blob/master/README.md

See https://github.com/nytimes/covid-19-data/blob/master/README.md#geographic-exceptions
for an explanation of geographic exceptions in this dataset.
}
\keyword{datasets}
